Migrated AIHal and WSI related routes to new folder
