+++ server/src/aiProjectAdmin/aiProjectSelectedWSImages.ts
